- make LockException available in API
