+++ bindings/storage-catches.i
@@ -19,6 +19,7 @@
 %exceptionclass storage::InvalidBlockSize;
 %exceptionclass storage::InvalidExtentSize;
 %exceptionclass storage::InvalidMountPointPath;
+%exceptionclass storage::LockException;
 %exceptionclass storage::LogicException;
 %exceptionclass storage::LvmLvNotFoundByLvName;
 %exceptionclass storage::LvmVgNotFoundByVgName;
@@ -233,7 +234,7 @@
 %catches(storage::DifferentBlockSizes) storage::Region::operator>(const Region &rhs) const;
 %catches(storage::DifferentBlockSizes) storage::Region::operator>=(const Region &rhs) const;
 %catches(storage::Exception) storage::Region::unused_regions(const std::vector< Region > &used_regions) const;
-%catches(storage::Exception) storage::Storage::Storage(const Environment &environment);
+%catches(storage::LockException, storage::Exception) storage::Storage::Storage(const Environment &environment);
 %catches(storage::Aborted, storage::Exception) storage::Storage::activate(const ActivateCallbacks *activate_callbacks) const;
 %catches(storage::Exception) storage::Storage::calculate_actiongraph();
 %catches(storage::Exception) storage::Storage::check(const CheckCallbacks *check_callbacks=nullptr) const;
