+++ doc/autodocs/doxygen.conf
@@ -31,8 +31,8 @@ FILE_PATTERNS += FilesystemUser.h Holder.h MdSubdevice.h MdUser.h Subdevice.h	\
 	User.h
 
 INPUT += ../../storage/Utils
-FILE_PATTERNS += Alignment.h Exception.h HumanString.h Logger.h Region.h	\
-	Remote.h Swig.h Topology.h
+FILE_PATTERNS += Alignment.h Exception.h HumanString.h Lock.h Logger.h		\
+	Region.h Remote.h Swig.h Topology.h
 
 JAVADOC_AUTOBRIEF = YES
